+++ _config.php
@@ -18,49 +18,20 @@ if (!defined('DC_CONTEXT_MODULE')) {
 $redir = empty($_REQUEST['redir']) ? 
     $list->getURL() . '#plugins' : $_REQUEST['redir'];
 
-# -- Combos of types of supported public pages --
-$types = [
-    __('home page')         => 'default',
-    __('post pages')        => 'post',
-    __('tags pages')        => 'tag',
-    __('archives pages')    => 'archive',
-    __('category pages')    => 'category',
-    __('entries feed')      => 'feed'
-];
-if ($core->plugins->moduleExists('muppet')) {
-    foreach(muppet::getPostTypes() as $k => $v) {
-        $types[sprintf(
-            __('"%s" pages from extension muppet'),
-            $v['name']
-        )] = $k;
-    }
-}
-
 # -- Get settings --
 $core->blog->settings->addNamespace('fac');
 $s = $core->blog->settings->fac;
 
-$fac_active             = (boolean) $s->fac_active;
-$fac_defaultfeedtitle   = (string) $s->fac_defaultfeedtitle;
-$fac_showfeeddesc       = (boolean) $s->fac_showfeeddesc;
-$fac_public_tpltypes    = @unserialize($s->fac_public_tpltypes);
-$fac_formats            = @unserialize($s->fac_formats);
+$fac_formats = @unserialize($s->fac_formats);
 
-if (!is_array($fac_public_tpltypes)) {
-    $fac_public_tpltypes = array();
-}
 if (!is_array($fac_formats)) {
-    $fac_formats = array();
+    $fac_formats = [];
 }
 
 # -- Set settings --
 if (!empty($_POST['save'])) {
     try {
-        $fac_active             = !empty($_POST['fac_active']);
-        $fac_defaultfeedtitle   = (string) $_POST['fac_defaultfeedtitle'];
-        $fac_showfeeddesc       = !empty($_POST['fac_showfeeddesc']);
-        $fac_public_tpltypes    = $_POST['fac_public_tpltypes'];
-        $fac_formats            = array();
+        $fac_formats = [];
 
         foreach($_POST['fac_formats'] as $uid => $f) {
             if (!empty($f['name'])) {
@@ -68,11 +39,16 @@ if (!empty($_POST['save'])) {
             }
         }
 
-        $s->put('fac_active',           $fac_active);
-        $s->put('fac_defaultfeedtitle', $fac_defaultfeedtitle);
-        $s->put('fac_showfeeddesc',     $fac_showfeeddesc);
-        $s->put('fac_public_tpltypes',  serialize($fac_public_tpltypes));
-        $s->put('fac_formats',          serialize($fac_formats));
+        // fix 2021.08.21 : formats are now global
+        $s->drop('fac_formats');
+        $s->put(
+            'fac_formats', 
+            serialize($fac_formats), 
+            'string', 
+            'Formats of feeds contents', 
+            true, 
+            true
+        );
 
         $core->blog->triggerBlog();
 
@@ -82,47 +58,12 @@ if (!empty($_POST['save'])) {
         http::redirect(
             $list->getURL('module=fac&conf=1&redir=' . $list->getRedir())
         );
-    }
-    catch (Exception $e) {
+    } catch (Exception $e) {
         $core->error->add($e->getMessage());
     }
 }
 
 # -- Display form --
-echo '
-<div class="fieldset">
-<h4>' . __('Activation') . '</h4>
-
-<p><label class="classic" for="fac_active">' .
-form::checkbox('fac_active', 1, $fac_active) .
-__('Enable plugin') . '</label></p>
-
-<h5>' . __('Show feed after content on:') . '</h5>';
-
-foreach($types as $k => $v) {
-    echo '
-    <p><label class="classic" for="fac_public_tpltypes' . $k . '">' .
-    form::checkbox(
-        array('fac_public_tpltypes[]', 'fac_public_tpltypes' . $k),
-        $v,
-        in_array($v, $fac_public_tpltypes)
-    ) . __($k) . '</label></p>';
-}
-echo '
-</div>
-
-<div class="fieldset">
-<h4>' . __('Feed') . '</h4>
-
-<p><label for="fac_defaultfeedtitle">' . __('Default title') . '</label>' .
-form::field('fac_defaultfeedtitle', 65, 255, $fac_defaultfeedtitle) . '</p>
-<p class="form-note">' . __('Use %T to insert title of feed.') . '</p>
-
-<p><label class="classic" for="fac_showfeeddesc">' .
-form::checkbox('fac_showfeeddesc', 1, $fac_showfeeddesc) .
-__('Show description of feed') . '</label></p>
-
-</div>';
 
 $i = 1;
 foreach($fac_formats as $uid => $f) {
@@ -139,10 +80,10 @@ foreach($fac_formats as $uid => $f) {
     <p><label for="fac_formats_' . $uid . '_name">' .
     __('Name:') . '</label>' .
     form::field(
-        array(
+        [
             'fac_formats[' . $uid . '][name]',
             'fac_formats_' . $uid . '_name'
-        ),
+        ],
         20,
         255,
         empty($f['name']) ? '' : $f['name'],
@@ -155,10 +96,10 @@ foreach($fac_formats as $uid => $f) {
     <p><label for="fac_formats_' . $uid . '_dateformat">' .
     __('Date format:') . '</label>' .
     form::field(
-        array(
+        [
             'fac_formats[' . $uid . '][dateformat]',
             'fac_formats_' . $uid . '_dateformat'
-        ),
+        ],
         20,
         255,
         empty($f['dateformat']) ? '' : $f['dateformat'],
@@ -171,10 +112,10 @@ foreach($fac_formats as $uid => $f) {
     <p><label for="fac_formats_' . $uid . '_lineslimit">' .
     __('Entries limit:') . '</label>' .
     form::field(
-        array(
+        [
             'fac_formats[' . $uid . '][lineslimit]',
             'fac_formats_' . $uid . '_lineslimit'
-        ),
+        ],
         5,
         4,
         empty($f['lineslimit']) ? '' : $f['lineslimit'],
@@ -189,10 +130,10 @@ foreach($fac_formats as $uid => $f) {
     <p><label for="fac_formats_' . $uid . '_linestitletext">' .
     __('Title format:') . '</label>' .
     form::field(
-        array(
+        [
             'fac_formats[' . $uid . '][linestitletext]',
             'fac_formats_' . $uid . '_linestitletext'
-        ),
+        ],
         20,
         255,
         empty($f['linestitletext']) ? '' : $f['linestitletext'],
@@ -210,10 +151,10 @@ foreach($fac_formats as $uid => $f) {
     <p><label for="fac_formats_' . $uid . '_linestitleover">' .
     __('Over title format:') . '</label>' .
     form::field(
-        array(
+        [
             'fac_formats[' . $uid . '][linestitleover]',
             'fac_formats_' . $uid . '_linestitleover'
-        ),
+        ],
         20,
         255,
         empty($f['linestitleover']) ? '' : $f['linestitleover'],
@@ -231,10 +172,10 @@ foreach($fac_formats as $uid => $f) {
     <p><label for="fac_formats_' . $uid . '_linestitlelength">' .
     __('Maximum length of title:') . '</label>' .
     form::field(
-        array(
+        [
             'fac_formats[' . $uid . '][linestitlelength]',
             'fac_formats_' . $uid . '_linestitlelength'
-        ),
+        ],
         5,
         4,
         empty($f['linestitlelength']) ? '' : $f['linestitlelength'],
@@ -248,10 +189,10 @@ foreach($fac_formats as $uid => $f) {
 
     <p><label for="fac_formats_' . $uid . '_showlinesdescription">' .
     form::checkbox(
-        array(
+        [
             'fac_formats[' . $uid . '][showlinesdescription]',
             'fac_formats_' . $uid . '_showlinesdescription'
-        ),
+        ],
         1,
         !empty($f['showlinesdescription'])
     ) .
@@ -259,10 +200,10 @@ foreach($fac_formats as $uid => $f) {
 
     <p><label for="fac_formats_' . $uid . '_linesdescriptionnohtml">' .
     form::checkbox(
-        array(
+        [
             'fac_formats[' . $uid . '][linesdescriptionnohtml]',
             'fac_formats_' . $uid . '_linesdescriptionnohtml'
-        ),
+        ],
         1,
         !empty($f['linesdescriptionnohtml'])
     ).
@@ -271,10 +212,10 @@ foreach($fac_formats as $uid => $f) {
     <p><label for="fac_formats_' . $uid . '_linesdescriptionlength">' .
     __('Maximum length of description:') . '</label>' .
     form::field(
-        array(
+        [
             'fac_formats[' . $uid . '][linesdescriptionlength]',
             'fac_formats_' . $uid . '_linesdescriptionlength'
-        ),
+        ],
         5,
         4,
         empty($f['linesdescriptionlength']) ? '' : $f['linesdescriptionlength'],
@@ -288,10 +229,10 @@ foreach($fac_formats as $uid => $f) {
 
     <p><label for="fac_formats_' . $uid . '_showlinescontent">' .
     form::checkbox(
-        array(
+        [
             'fac_formats[' . $uid . '][showlinescontent]',
             'fac_formats_' . $uid . '_showlinescontent'
-        ),
+        ],
         1,
         !empty($f['showlinescontent'])
     ) .
@@ -299,10 +240,10 @@ foreach($fac_formats as $uid => $f) {
 
     <p><label for="fac_formats_' . $uid . '_linescontentnohtml">' .
     form::checkbox(
-        array(
+        [
             'fac_formats[' . $uid . '][linescontentnohtml]',
             'fac_formats_' . $uid . '_linescontentnohtml'
-        ),
+        ],
         1,
         !empty($f['linescontentnohtml'])
     ) .
@@ -311,10 +252,10 @@ foreach($fac_formats as $uid => $f) {
     <p><label for="fac_formats_' . $uid . '_linescontentlength">' .
     __('Maximum length of content:') . '</label>' .
     form::field(
-        array(
+        [
             'fac_formats[' . $uid . '][linescontentlength]',
             'fac_formats_' . $uid . '_linescontentlength'
-        ),
+        ],
         5,
         4,
         empty($f['linescontentlength']) ? '' : $f['linescontentlength'],
@@ -341,10 +282,10 @@ echo '
 <p><label for="fac_formats_' . $uid . '_name">' .
 __('Name:') . '</label>' .
 form::field(
-    array(
+    [
         'fac_formats[' . $uid . '][name]',
         'fac_formats_' . $uid . '_name'
-    ),
+    ],
     20,
     255,
     '',
@@ -357,10 +298,10 @@ __('In order to remove a format, leave its name empty.') .
 <p><label for="fac_formats_' . $uid . '_dateformat">' .
 __('Date format:') . '</label>' .
 form::field(
-    array(
+    [
         'fac_formats[' . $uid . '][dateformat]',
         'fac_formats_' . $uid . '_dateformat'
-    ),
+    ],
     20,
     255,
     '',
@@ -373,10 +314,10 @@ __('Use date format of Dotclear or leave empty to use default date format of blo
 <p><label for="fac_formats_' . $uid . '_lineslimit">' .
 __('Entries limit:') . '</label>' .
 form::field(
-    array(
+    [
         'fac_formats[' . $uid . '][lineslimit]',
         'fac_formats_' . $uid . '_lineslimit'
-    ),
+    ],
     5,
     4,
     5,
@@ -391,10 +332,10 @@ __('Leave lengh empty for no limit.') .
 <p><label for="fac_formats_' . $uid . '_linestitletext">' .
 __('Title format:') . '</label>' .
 form::field(
-    array(
+    [
         'fac_formats[' . $uid . '][linestitletext]',
         'fac_formats_' . $uid . '_linestitletext'
-    ),
+    ],
     20,
     255,
     '%T',
@@ -412,10 +353,10 @@ __('Format can be:') .
 <p><label for="fac_formats_' . $uid . '_linestitleover">' .
 __('Over title format:')  .  '</label>' .
 form::field(
-    array(
+    [
         'fac_formats[' . $uid . '][linestitleover]',
         'fac_formats_' . $uid . '_linestitleover'
-    ),
+    ],
     20,
     255,
     '%D',
@@ -433,10 +374,10 @@ __('Format can be:') .
 <p><label for="fac_formats_' . $uid . '_linestitlelength">' .
 __('Maximum length of title:') . '</label>' .
 form::field(
-    array(
+    [
         'fac_formats[' . $uid . '][linestitlelength]',
         'fac_formats_' . $uid . '_linestitlelength'
-    ),
+    ],
     5,
     4,
     150,
@@ -450,10 +391,10 @@ __('Leave lengh empty for no limit.') .
 
 <p><label for="fac_formats_' . $uid . '_showlinesdescription">' .
 form::checkbox(
-    array(
+    [
         'fac_formats[' . $uid . '][showlinesdescription]',
         'fac_formats_' . $uid . '_showlinesdescription'
-    ),
+    ],
     1,
     0
 ) .
@@ -461,10 +402,10 @@ __('Show description of entries') . '</label></p>
 
 <p><label for="fac_formats_' . $uid . '_linesdescriptionnohtml">' .
 form::checkbox(
-    array(
+    [
         'fac_formats[' . $uid . '][linesdescriptionnohtml]',
         'fac_formats_' . $uid . '_linesdescriptionnohtml'
-    ),
+    ],
     1,
     1
 ) .
@@ -473,10 +414,10 @@ __('Remove html of description') . '</label></p>
 <p><label for="fac_formats_' . $uid . '_linesdescriptionlength">' .
 __('Maximum length of description:') . '</label>' .
 form::field(
-    array(
+    [
         'fac_formats[' . $uid . '][linesdescriptionlength]',
         'fac_formats_' . $uid . '_linesdescriptionlength'
-    ),
+    ],
     5,
     4,
     350,
@@ -490,10 +431,10 @@ __('Leave lengh empty for no limit.') .
 
 <p><label for="fac_formats_' . $uid . '_showlinescontent">' .
 form::checkbox(
-    array(
+    [
         'fac_formats[' . $uid . '][showlinescontent]',
         'fac_formats_' . $uid . '_showlinescontent'
-    ),
+    ],
     1,
     0
 ) .
@@ -501,10 +442,10 @@ __('Show content of entries') . '</label></p>
 
 <p><label for="fac_formats_' . $uid . '_linescontentnohtml">' .
 form::checkbox(
-    array(
+    [
         'fac_formats[' . $uid . '][linescontentnohtml]',
         'fac_formats_' . $uid . '_linescontentnohtml'
-    ),
+    ],
     1,
     1
 ) .
@@ -513,10 +454,10 @@ __('Remove html of content') . '</label></p>
 <p><label for="fac_formats_' . $uid . '_linescontentlength">' .
 __('Maximum length of content:') . '</label>' .
 form::field(
-    array(
+    [
         'fac_formats[' . $uid . '][linescontentlength]',
         'fac_formats_' . $uid . '_linescontentlength'
-    ),
+    ],
     5,
     4,
     350,
@@ -558,3 +499,5 @@ __('To add feed to an entry edit this entry and put in sidebar the url of the fe
 </div>
 
 </div>';
+
+dcPage::helpBlock('fac');